fix(space-bar): stage-3 review fixes
Two-agent review on a7a81ad:

- Vertical preview no longer overflows its 96 pt canvas (H1 —
  the default edge was the broken case): tighter vertical
  thickness/cell/gap budget, two one-glyph mock items instead
  of three, clipShape belt-and-braces; mock content split to
  +Mock.swift.
- Same-edge row and preview stand-in now use one shared core
  predicate (TilingSettings.spaceBarSharesEdgeWithAppBar):
  true only while at least one ENABLED layout App Bar's
  resolved edge — per-layout overrides included — matches the
  Space Bar's. Both bars off no longer claims a coexistence
  that doesn't exist. Unit-tested.
- mutedBadgeAlpha / frontDividerAlpha re-homed as public
  statics on SpaceBarStyle; runtime views and the preview now
  share one derivation instead of mirrored literals.
- Space Bar section retitled "Space Bar style" — distinct from
  the App Bar's "Global style", so the search caption is
  unambiguous and the index entry reachable; the switch-landing
  limit (hits open the App Bar editor one click away) is
  recorded in SidebarSearch, deferred to #277.
- Plan §10 design-decisions rows added: two-accent model, v1
  fixed content, icon-only identifier + fallbacks.
- Copy-appearance comment rewords the inert `enabled`
  exclusion; Sections/AppBarSection.swift renamed to
  BarsSection.swift.

Part of #293 (stage 3).
