+++ Sources/KiwiDesk/Settings/Tabs/SpaceBarPreviewStrip.swift
@@ -4,20 +4,23 @@ import SwiftUI
 /// A static, in-window mock of the Space Bar (#293) — the
 /// `AppBarPreviewStrip` idiom: literal configured hex colors
 /// over a neutral backdrop, pure SwiftUI, no runtime reads.
-/// Three mock Space items exercise the states in one glance:
-/// an inactive space with two glyphs (one wearing a count
-/// badge, muted), the active space showing both accents (its
-/// identifier + a focused glyph), and an empty space
-/// (identifier only). Edge-aware: rotates into a column for
-/// left/right. When both enabled bars share an edge, a thin
+/// Edge-aware: a row for top/bottom, a column for left/right.
+/// When both enabled bars share an edge (`sameEdge`, resolved
+/// by `TilingSettings.spaceBarSharesEdgeWithAppBar`), a muted
 /// App Bar stand-in draws on the window-facing side so the
 /// stacking order is seen, not described.
 ///
-/// A schematic, not a pixel-mirror — the rendered truth lives
-/// in `SpaceBarItemView` (KiwiDeskCore); keep the two in sync.
+/// Horizontal shows three mock Spaces (inactive with a muted
+/// count badge, active with both accents, empty); the vertical
+/// budget is tighter — 96 pt must hold whole items — so it
+/// shows the two state-bearing items with one glyph each
+/// (`AppBarPreviewStrip.verticalSlotLength` precedent). Mock
+/// content lives in SpaceBarPreviewStrip+Mock.swift; keep in
+/// sync with `SpaceBarItemView` (the rendered truth).
 struct SpaceBarPreviewStrip: View {
     let style: SpaceBarStyle
     let appBar: AppBarStyle
+    let sameEdge: Bool
 
     var body: some View {
         VStack(spacing: 4) {
@@ -27,6 +30,9 @@ struct SpaceBarPreviewStrip: View {
                 content
             }
             .frame(height: 96)
+            // Belt-and-braces: a mock must never spill over
+            // the caption and controls below.
+            .clipShape(RoundedRectangle(cornerRadius: 6))
             .animation(LayoutSchematic.damping, value: style)
             .accessibilityElement()
             .accessibilityLabel(axLabel)
@@ -38,10 +44,6 @@ struct SpaceBarPreviewStrip: View {
     }
 
     // MARK: - Composition
-
-    private var sameEdge: Bool {
-        style.enabled && style.edge == appBar.edge
-    }
 
     /// The strip plus, when stacked, the App Bar stand-in on
     /// the window-facing side of the shared edge.
@@ -80,7 +82,7 @@ struct SpaceBarPreviewStrip: View {
         stack {
             item(.inactive)
             item(.active)
-            item(.empty)
+            if style.edge.isHorizontal { item(.empty) }
         }
         .padding(4)
         .background(
@@ -108,151 +110,33 @@ struct SpaceBarPreviewStrip: View {
         }
     }
 
-    // MARK: - Mock items
-
-    private enum MockState { case inactive, active, empty }
-
-    @ViewBuilder private func item(
-        _ state: MockState
-    ) -> some View {
-        let cells = cellViews(state)
-        Group {
-            if style.edge.isHorizontal {
-                HStack(spacing: 2) { cells }
-            } else {
-                VStack(spacing: 2) { cells }
-            }
-        }
-        .padding(3)
-        .background(
-            RoundedRectangle(
-                cornerRadius: style.tabBackground == .boxed
-                    ? corner : 0
-            )
-            .fill(itemBox(state))
-        )
-        .overlay(indicator(state))
-    }
-
-    @ViewBuilder private func cellViews(
-        _ state: MockState
-    ) -> some View {
-        // Identifier leads; the active item shows the
-        // two-accent pair, the inactive one a muted badge.
-        identifierCell(state)
-        switch state {
-        case .inactive:
-            glyphCell(tint: mutedColor, badge: true)
-            glyphCell(tint: mutedColor, badge: false)
-        case .active:
-            glyphCell(
-                tint: color(style.activeTextColor),
-                badge: false
-            )
-            glyphCell(
-                tint: color(style.focusedItemColor),
-                badge: false
-            )
-        case .empty:
-            EmptyView()
-        }
-    }
-
-    private func identifierCell(_ state: MockState) -> some View {
-        Image(
-            systemName: state == .active
-                ? "2.square"
-                : state == .empty
-                    ? "3.square" : "1.square"
-        )
-        .font(.system(size: font))
-        .foregroundStyle(
-            state == .active
-                ? color(style.activeTextColor) : mutedColor
-        )
-        .frame(width: cell, height: cell)
-    }
-
-    private func glyphCell(
-        tint: Color,
-        badge: Bool
-    ) -> some View {
-        Circle()
-            .fill(tint)
-            .frame(width: cell * 0.55, height: cell * 0.55)
-            .frame(width: cell, height: cell)
-            .overlay(alignment: .topTrailing) {
-                if badge { badgeDot }
-            }
-    }
-
-    /// The count badge, muted on the inactive mock item —
-    /// the same derivation the runtime uses.
-    private var badgeDot: some View {
-        Text("2")
-            .font(.system(size: 6, weight: .bold))
-            .foregroundStyle(color(style.textColor))
-            .padding(2)
-            .background(
-                Circle().fill(
-                    color(style.textColor).opacity(0.3)
-                )
-            )
-    }
-
-    @ViewBuilder private func indicator(
-        _ state: MockState
-    ) -> some View {
-        if state == .active {
-            switch style.activeIndicator {
-            case .ring:
-                RoundedRectangle(cornerRadius: corner)
-                    .strokeBorder(
-                        color(style.highlightColor),
-                        lineWidth: 2
-                    )
-            case .edgeMark:
-                edgeMark
-            case .gap:
-                EmptyView()
-            }
-        }
-    }
-
-    @ViewBuilder private var edgeMark: some View {
-        if style.edge.isHorizontal {
-            Rectangle()
-                .fill(color(style.highlightColor))
-                .frame(height: 2)
-                .frame(
-                    maxHeight: .infinity,
-                    alignment: style.edge == .top
-                        ? .bottom : .top
-                )
-        } else {
-            Rectangle()
-                .fill(color(style.highlightColor))
-                .frame(width: 2)
-                .frame(
-                    maxWidth: .infinity,
-                    alignment: style.edge == .left
-                        ? .trailing : .leading
-                )
-        }
-    }
-
     // MARK: - Metrics
 
-    private var thickness: CGFloat {
-        scale(style.thickness, from: 8...80, to: 16...40)
+    /// Cross-axis depth. The vertical range is tighter so two
+    /// whole items + the strip padding stay inside the 96 pt
+    /// canvas at every thickness.
+    var thickness: CGFloat {
+        style.edge.isHorizontal
+            ? scale(style.thickness, from: 8...80, to: 16...40)
+            : scale(style.thickness, from: 8...80, to: 14...24)
     }
-    private var cell: CGFloat { thickness * 0.62 }
-    private var gap: CGFloat {
+
+    /// Square glyph cell. Vertical budget: two items of
+    /// (2 cells + 2 spacing + 6 padding) + one gap + 8 strip
+    /// padding ≤ 96 → cell ≤ 13 at the maxima.
+    var cell: CGFloat {
+        style.edge.isHorizontal
+            ? thickness * 0.62
+            : min(thickness * 0.62, 13)
+    }
+
+    var gap: CGFloat {
         style.edge.isHorizontal
             ? scale(style.itemGap, from: 0...40, to: 0...14)
-            : scale(style.itemGap, from: 0...40, to: 0...6)
+            : scale(style.itemGap, from: 0...40, to: 0...4)
     }
-    private var corner: CGFloat {
+
+    var corner: CGFloat {
         min(
             style.resolvedCornerRadius(
                 forThickness: thickness
@@ -260,23 +144,15 @@ struct SpaceBarPreviewStrip: View {
             thickness / 2
         )
     }
-    private var font: CGFloat {
+
+    var font: CGFloat {
         let base =
             style.fontSize > 0
             ? style.fontSize : thickness * 0.4
-        return min(base, thickness * 0.5)
+        return min(base, thickness * 0.5, cell)
     }
-    private var mutedColor: Color { color(style.textColor) }
 
-    private func itemBox(_ state: MockState) -> Color {
-        guard style.tabBackground == .boxed else {
-            return .clear
-        }
-        return color(
-            state == .active
-                ? style.activeBoxColor : style.boxColor
-        )
-    }
+    var mutedColor: Color { color(style.textColor) }
 
     private func scale(
         _ value: CGFloat,
@@ -291,7 +167,7 @@ struct SpaceBarPreviewStrip: View {
             * (dst.upperBound - dst.lowerBound)
     }
 
-    private func color(_ hex: String) -> Color {
+    func color(_ hex: String) -> Color {
         Color(kiwiHex: hex)
     }
