+++ Sources/KiwiDeskCore/Resources/Locales/en.json
@@ -623,7 +623,7 @@
   "space_bar.font_size.auto": "Auto font size",
   "space_bar.front_app.ax": "Front app: %1$@",
   "space_bar.global_style.caption": "One bar per display, every layout. All settings are global.",
-  "space_bar.global_style.title": "Global style",
+  "space_bar.global_style.title": "Space Bar style",
   "space_bar.hide_empty": "Hide empty Spaces",
   "space_bar.hide_empty.help": "Spaces with no windows are hidden from the bar, except the Space you're currently on. Use a shortcut to jump to a hidden Space.",
   "space_bar.icon_source.help": "How app glyphs are drawn. Glyphs shows a monochrome symbol colored by the bar's text colors; apps without a symbol keep their app icon.",
